backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.86.0
anthropic>=0.30.0
pydantic>=2.12.2
pydantic[email]>=2.12.2
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.29.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
